refactor: move extract logic into separate module
